refactor: new name for vec mul
