+++ src/shader_tests/mod.rs
@@ -1,1 +1,2 @@
-mod dot;
+mod vec_mul;
+mod sum_reduce;
